+++ source/NewSOFIKernels.cpp
@@ -54,7 +54,7 @@ bool ComparePixelCombinations::operator() (const PixelCombination& comb1, const 
     if (comb1.size() > comb2.size())
         return false;
     
-    for (int i = 0; i < comb1.size(); ++i) {
+    for (size_t i = 0; i < comb1.size(); ++i) {
         if (comb1[i] == comb2[i])
             continue;
         return comb1[i] < comb2[i];
